Moved component styling into their respective directories
