+++ src/components/Categories/CategoryBox.scss
@@ -1,0 +1,16 @@
+.categories {
+    &__icon {
+        width: 7rem;
+        padding: 0.4rem;
+        margin: 0.6rem auto;
+        border-radius: 1rem;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+    }
+
+    &__text {
+        margin: 0.4rem auto;
+        color: #fff;
+    }
+}
