+++ src/components/TransactionForm/TransactionForm.scss
@@ -1,0 +1,14 @@
+.form {
+    &__heading {
+        font-size: 2rem;
+        text-align: center;
+        margin: 2rem auto;
+    }
+}
+
+.categories {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+}
